gl_sdl2: Huge performance boost
SDL_RenderDrawPoint is a horrible solution for writing huge
amounts of data. Let's allocate a framebuffer to avoid excessive
shared library calls and access it directly. To display the frame we
just need to copy it to GPU memory with SDL_UpdateTexture.

From 70-100 FPS I got 1100-1200 in triangle1... (without usleep of
course).