gl: Avoid z-fighting
When we draw two surfaces sharing the same vertex positions
(examples/triangle2), we need to decide who is in front. To eliminate
this ambiguity, let's add 2 more iterations to zstep, ignoring
the ambiguous initial and final Z value. The first iteration
is ignored by starting at z0 + zstep, and the last by not being
reached by the loop.